reactor: navbar into separate shared component
Moved the navigation sidebar logic from Default.tsx to a new shared/Navbar.tsx component for better modularity and code organization. Updated Default.tsx to use the new Navbar component and removed related logic and imports.
